# eliminate the use of Mirror