--- a/.swiftpm/xcode/xcshareddata/xcschemes/CasePaths.xcscheme
+++ b/.swiftpm/xcode/xcshareddata/xcschemes/CasePaths.xcscheme
@@ -27,6 +27,7 @@
       selectedDebuggerIdentifier = "Xcode.DebuggerFoundation.Debugger.LLDB"
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       shouldUseLaunchSchemeArgsEnv = "YES"
+      disableMainThreadChecker = "YES"
       codeCoverageEnabled = "YES">
       <Testables>
          <TestableReference