--- a/Package@swift-5.1.swift
+++ b/Package@swift-5.1.swift
@@ -12,15 +12,11 @@ let package = Package(
   ],
   targets: [
     .target(
-      name: "CasePaths",
-      dependencies: ["CCasePaths"]
+      name: "CasePaths"
     ),
     .testTarget(
       name: "CasePathsTests",
       dependencies: ["CasePaths"]
-    ),
-    .target(
-      name: "CCasePaths"
     ),
   ]
 )