--- a/Sources/CCasePaths/Builtin_NativeObject_metadata.c
+++ b/Sources/CCasePaths/Builtin_NativeObject_metadata.c
@@ -1,8 +1,0 @@
-#include "Builtin_NativeObject_metadata.h"
-
-// This is the mangled name of Builtin.NativeObject's metadata.
-extern void $sBoN;
-
-void const *getBuiltinNativeObjectFullMetadata() {
-    return &$sBoN;
-}