This branch is not ready for merging. I'm creating this pull request only to make it available for discussion.

I found example uses of `swift_getTypeByMangledNameInContext` in [the Reflect package](https://github.com/wickwirew/Runtime) and have made some progress toward using it to replace the use of `Mirror`.

In its current state, this branch has several failing tests, all related to associated value labels. I've outlined what's wrong and what needs to be done in the commit message and code comments.