@mayoff 2021-07-10
The new version ([7cab867](https://github.com/pointfreeco/swift-case-paths/pull/37/commits/7cab86738ddb2be55561a91241a54085a534981b)) fixes the existential bug and passes all the tests (including some new ones).

Code coverage of `EnumReflection.swift` indicates the tests cover everything except an “impossible” case where some metadata is missing, and a mangled type name encoding that I don't know how to trigger.

It also runs faster.

Main branch ([90e4e67e](https://github.com/pointfreeco/swift-case-paths/commit/90e4e67ee18f72f2b1251211214ffc1fba60e540)):

```
name               time       std        iterations
---------------------------------------------------
Success.Manual      34.000 ns ± 146.77 %    1000000
Success.Reflection 151.000 ns ± 121.01 %    1000000
Failure.Manual      35.000 ns ± 231.80 %    1000000
Failure.Reflection  65.000 ns ± 191.00 %    1000000
```

New version:

```
name               time       std        iterations
---------------------------------------------------
Success.Manual      33.000 ns ± 226.73 %    1000000
Success.Reflection 128.000 ns ±  85.30 %    1000000
Failure.Manual      37.000 ns ± 149.48 %    1000000
Failure.Reflection  73.000 ns ± 140.36 %    1000000
```