@mayoff 2021-07-10
Surprisingly, unwrapping just the cached tag first to reject non-matching values (instead of unwrapping the whole cached extractor) consistently saves a few seconds in the Failure.Reflection benchmark.

```
name               time       std        iterations
---------------------------------------------------
Success.Manual      34.000 ns ± 261.66 %    1000000
Success.Reflection 135.000 ns ± 116.75 %    1000000
Failure.Manual      36.000 ns ± 221.68 %    1000000
Failure.Reflection  66.000 ns ± 174.90 %    1000000
```